[SimpleRSS] type 'video/mp4' and 'audio/mpeg' added

+++ simplerss/src/plugin.py
@@ -17,7 +17,7 @@ from xml.etree.ElementTree import tostring, parse, fromstring
 # self.feeds: [(<Plugins.Extensions.SimpleRSS.plugin.BaseFeed object at 0xa688def0>,), (<Plugins.Extensions.SimpleRSS.plugin.UniversalFeed object at 0xa6890030>,), (<Plugins.Extensions.SimpleRSS.plugin.UniversalFeed object at 0xa68900b0>,), (<Plugins.Extensions.SimpleRSS.plugin.UniversalFeed object at 0xa68900d0>,), (<Plugins.Extensions.SimpleRSS.plugin.UniversalFeed object at 0xa68900f0>,), (<Plugins.Extensions.SimpleRSS.plugin.UniversalFeed object at 0xa6890110>,)]
 
 # ENIGMA IMPORTS
-from enigma import getDesktop, eTimer, BT_SCALE, BT_KEEP_ASPECT_RATIO, BT_HALIGN_CENTER, BT_VALIGN_CENTER
+from enigma import getDesktop, eTimer, eServiceReference, BT_SCALE, BT_KEEP_ASPECT_RATIO, BT_HALIGN_CENTER, BT_VALIGN_CENTER
 from Components.ActionMap import ActionMap
 from Components.config import config, getConfigListEntry, ConfigSubsection, ConfigSubList, ConfigEnableDisable, ConfigNumber, ConfigSelectionNumber, ConfigText, ConfigSelection
 from Components.ConfigList import ConfigListScreen
@@ -30,6 +30,7 @@ from Plugins.Extensions.PicturePlayer import ui
 from Plugins.Plugin import PluginDescriptor
 from Plugins.SystemPlugins.Toolkit.TagStrip import strip, strip_readable
 from Screens.ChoiceBox import ChoiceBox
+from Screens.InfoBar import MoviePlayer
 from Screens.MessageBox import MessageBox
 from Screens.Screen import Screen
 from Tools import Notifications
@@ -54,7 +55,10 @@ NOTIFICATIONID = 'SimpleRSSUpdateNotification'
 NS_RDF = "{http://www.w3.org/1999/02/22-rdf-syntax-ns#}"
 NS_RSS_09 = "{http://my.netscape.com/rdf/simple/0.9/}"
 NS_RSS_10 = "{http://purl.org/rss/1.0/}"
-
+MIMEIMAGES = ["image/jpg", "image/jpeg", "image/png", "image/gif", "image/webp"]
+MIMEVIDEOS = ["video/mp4"]
+MIMEAUDIOS = ["audio/mpeg"]
+EXT2MIME = {".jpg": "image/jpg", ".jpeg": "image/jpeg", ".png": "image/png", ".gif": "image/gif", ".webp": "image/webp", "mp4": "video/mp4", "mp3": "audio/mpeg"}
 # Initialize Configuration
 config.plugins.simpleRSS = ConfigSubsection()
 simpleRSS = config.plugins.simpleRSS
@@ -144,35 +148,38 @@ def downloadPicfile(url, picfile, resize=None, fixedname=None, callback=None):
 				picfile = fixedname
 		except exceptions.RequestException as err:
 			print("[%s] ERROR in module 'downloadPicfile': '%s" % (MODULE_NAME, str(err)))
-		with open(picfile, "wb") as f:
-			f.write(response.content)
-		isnotpng = not picfile.endswith(".png")
-		if resize or isnotpng:
-			try:  # in case PIL cannot handle picture format
-				img = Image.open(picfile)
-				if resize:
-					img.thumbnail(resize, Image.LANCZOS)  # resize, keeping aspect ratio and antialiasing
-				pngfile = "%s.png" % picfile.split(".")[0]
-				img.save(pngfile, format="png", lossless=True)  # forced save as PNG because some boxes (e.g. HD51) wrongly display JPGs transparently
-				img.close()
-			except Exception as err:
-				pngfile = join(TEMPPATH, NOPIC)
-				print("[%s] ERROR in module 'downloadPicfile': unsupported or invalid picture format for '%s': %s" % (MODULE_NAME, str(err), pngfile))
-			if isnotpng:
-				remove(picfile)
-		else:
-			pngfile = picfile
-		if callback:
-			callback()
+			return
+		if exists(picfile[:picfile.rfind("/")]):
+			with open(picfile, "wb") as f:
+				f.write(response.content)
+			isnotpng = not picfile.endswith(".png")
+			if resize or isnotpng:
+				try:  # in case PIL cannot handle picture format
+					img = Image.open(picfile)
+					if resize:
+						img.thumbnail(resize, Image.LANCZOS)  # resize, keeping aspect ratio and antialiasing
+					pngfile = "%s.png" % picfile[:picfile.rfind(".")]
+					img.save(pngfile, format="png", lossless=True)  # forced save as PNG because some boxes (e.g. HD51) wrongly display JPGs transparently
+					img.close()
+				except Exception as err:
+					pngfile = join(TEMPPATH, NOPIC)
+					print("[%s] ERROR in module 'downloadPicfile': unsupported or invalid picture format for '%s': %s" % (MODULE_NAME, str(err), pngfile))
+				if isnotpng and exists(picfile):
+					remove(picfile)
+			else:
+				pngfile = picfile
+			if callback:
+				callback()
 	else:
 		print("[%s] ERROR in module 'downloadPicfile': missing link or picfile." % MODULE_NAME)
 
 
 def url2filename(url, forcepng=False):
-	for itype in [".jpg", ".jpeg", ".png", ".gif", ".webp"]:
-		if itype in url:
-			return "%s.png" % hash(url) if forcepng else ("%s%s" % (hash(url), itype)).replace(".jpeg", ".jpg")
-	return ""
+	return "%s.png" % hash(url) if forcepng else ("%s%s" % (hash(url), url[url.rfind("."):])).replace(".jpeg", ".jpg")
+
+
+def isExtensionSupported(name, supportlist=MIMEIMAGES):
+    return name[name.rfind("."):] in [".%s" % type[type.rfind("/") + 1:] for type in supportlist]  # extension supported?
 
 
 class RSS_TickerView(Screen):
@@ -431,12 +438,11 @@ class RSSBaseView(Screen):  # Base Screen for all Screens used in SimpleRSS
 			self.pollDialog = self.session.open(MessageBox, _("Update is being done in Background.\nContents will automatically be updated when it's done."), type=MessageBox.TYPE_INFO, timeout=3)
 
 	def findEnclosure(self, enclosures):
-		if enclosures:
-			urllist = []
-			for enclosure in enclosures:
-				if enclosure[1] in ["image/jpg", "image/jpeg", "image/png", "image/gif", "image/webp"]:
-					urllist.append(enclosure[0].strip())
-			return urllist
+		enclist = []
+		for enclosure in enclosures:
+			if enclosure[1] in MIMEIMAGES + MIMEVIDEOS + MIMEAUDIOS:
+				enclist.append((enclosure[0].strip(), enclosure[1]))
+		return enclist
 
 
 class RSS_EntryView(RSSBaseView):  # Shows a RSS Item
@@ -448,16 +454,28 @@ class RSS_EntryView(RSSBaseView):  # Shows a RSS Item
 		self.cur_idx = cur_idx
 		self.entries = entries
 		self.feedpng = feedLogo
+		self.enclist = []
 		self.pngfile = join(TEMPPATH, "entrypic.png")
 		self.picsize = (382, 214) if getDesktop(0).size().width() > 1300 else (255, 143)
 		self.skin = readSkin("RSS_EntryView")
 		Screen.__init__(self, session, self.skin)
 		self["title"] = StaticText(_("Simple RSS Reader EntryView"))
 		self["info"] = StaticText(_("Entry %s/%s") % (cur_idx + 1, entries)) if cur_idx is not None and entries is not None else StaticText()
-		self["content"] = ScrollLabel(''.join((data[0], '\n\n', data[2], '\n\n', str(len(data[3])), ' ', _("Enclosures")))) if data else ScrollLabel()
 		self["feedlogo"] = Pixmap()
 		self["picture"] = Pixmap()
-		self["actions"] = ActionMap(["OkCancelActions"], {"ok": self.selectEnclosure, "cancel": self.__close}, -1)
+		self["enctext"] = StaticText(_("Enclosures"))
+		self["enclist"] = List([])
+		self["content"] = ScrollLabel(''.join((data[0], '\n\n', data[2], '\n\n', str(len(data[3])), ' ', _("Enclosures")))) if data else ScrollLabel()
+		self["actions"] = ActionMap(["OkCancelActions", "DirectionActions", "ChannelSelectBaseActions"],
+						{"ok": self.showEnclosure,
+						"cancel": self.__close,
+						"up": self.keyPageUp,
+						"down": self.keyPageDown,
+						"right": self.keyPageDown,
+						"left": self.keyPageUp,
+						"nextBouquet": self.nextEnclosure,
+						"prevBouquet": self.prevEnclosure
+						}, -1)
 		self.onLayoutFinish.append(self.updateTitle)
 
 	def updateTitle(self):
@@ -472,13 +490,18 @@ class RSS_EntryView(RSSBaseView):  # Shows a RSS Item
 		self.setTitle(_("Simple RSS Reader: %s") % (self.feedTitle))
 		self["info"].text = _("Entry %s/%s") % (self.cur_idx + 1, self.entries) if self.cur_idx is not None and self.entries else ""
 		data = self.data
-		self["content"].setText(''.join((data[0], '\n\n', data[2], '\n\n', str(len(data[3])), ' ', _("Enclosures"))) if data else _("No such Item."))
+		self["content"].setText(''.join((data[0], '\n\n', data[2])) if data else _("No such Item."))
 		if data:
-			urllist = RSSBaseView.findEnclosure(self, data[3])
-			if urllist:
-				picfile = join(TEMPPATH, url2filename(urllist[0]))
-				if not exists(self.pngfile):
-					callInThread(downloadPicfile, urllist[0], picfile, resize=self.picsize, fixedname=self.pngfile, callback=self.refreshPic)
+			enclist = RSSBaseView.findEnclosure(self, data[3])
+			for enclosure in enclist:
+				if isExtensionSupported(enclosure[0]):
+					picfile = join(TEMPPATH, url2filename(enclosure[0]))
+					if isExtensionSupported(picfile) and not exists(self.pngfile):
+						callInThread(downloadPicfile, enclosure[0], picfile, resize=self.picsize, fixedname=self.pngfile, callback=self.refreshPic)
+			if not enclist:
+				enclist = [(None, _("{no attachments found}"))]
+			self["enclist"].setList([enclosure[1] for enclosure in enclist])
+			self.enclist = enclist
 		self.updateInfo()
 
 	def updateInfo(self):
@@ -493,24 +516,39 @@ class RSS_EntryView(RSSBaseView):  # Shows a RSS Item
 		self["picture"].instance.setPixmapFromFile(pngfile)
 		self["picture"].show()
 
-	def selectEnclosure(self):
-		if self.data:
-			urllist = RSSBaseView.findEnclosure(self, self.data[3])
-			if urllist:
-				for url in urllist:
-					self.filename = join(TEMPPATH, "full_%s" % url2filename(url))
-					if exists(self.filename):
-						self.showFullpic()
-					else:
-						callInThread(downloadPicfile, url, self.filename, callback=self.showFullpic)
+	def showEnclosure(self):
+		currindex = self["enclist"].getSelectedIndex()
+		if currindex < len(self.enclist):
+			if self.enclist[currindex][1] in MIMEIMAGES:
+				self.filename = join(TEMPPATH, "full_%s" % url2filename(self.enclist[currindex][0]))
+				if exists(self.filename):
+					self.showFullpic()
+				else:
+					callInThread(downloadPicfile, self.enclist[currindex][0], self.filename, callback=self.showFullpic)
+			elif self.enclist[currindex][1] in MIMEVIDEOS + MIMEAUDIOS:
+				self.session.open(MoviePlayer, eServiceReference(4097, 0, self.enclist[currindex][0]))
+			elif self.enclist[0][0]:
+				self.session.open(MessageBox, _("Sorry, this attachment cannot be opened: unsupported MIME type"), type=MessageBox.TYPE_ERROR, timeout=5)
 
 	def showFullpic(self):
-		pngfile = join(TEMPPATH, "%s.png" % self.filename.split(".")[0])
+		pngfile = join(TEMPPATH, "%s.png" % self.filename[:self.filename.rfind(".")])
 		piclist = []
 		if exists(pngfile):
 			piclist.append(((pngfile, False), None))
 		if piclist:
 			self.session.open(ui.Pic_Full_View, piclist, 0, TEMPPATH)
+
+	def keyPageUp(self):
+		self["content"].pageUp()
+
+	def keyPageDown(self):
+		self["content"].pageDown()
+
+	def nextEnclosure(self):
+		self["enclist"].up()
+
+	def prevEnclosure(self):
+		self["enclist"].down()
 
 	def __close(self):
 		if exists(self.pngfile):
@@ -580,10 +618,12 @@ class RSS_FeedView(RSSBaseView):  # Shows a RSS-Feed
 		self.buildSkinList()
 		self.updateInfo()
 		if self.feed:
+			logopng = join(TEMPPATH, NEWSLOGO) if self.feed.title == _("New Items") else join(TEMPPATH, NOLOGO)
 			if self.feed.logoUrl:
-				logopng = join(TEMPPATH, url2filename(self.feed.logoUrl, forcepng=True))
-			else:
-				logopng = join(TEMPPATH, NEWSLOGO) if self.feed.title == _("New Items") else join(TEMPPATH, NOLOGO)
+				for url in self.feed.logoUrl:  # catch first image found in list
+					if isExtensionSupported(url):
+						logopng = join(TEMPPATH, url2filename(url, forcepng=True))
+						break
 			if exists(logopng):
 				self["feedlogo"].instance.setPixmapScaleFlags(BT_SCALE | BT_KEEP_ASPECT_RATIO | BT_HALIGN_CENTER | BT_VALIGN_CENTER)
 				self["feedlogo"].instance.setPixmapFromFile(logopng)
@@ -624,15 +664,17 @@ class RSS_FeedView(RSSBaseView):  # Shows a RSS-Feed
 		skinlist = []
 		if self.feed:
 			for content in self.feed.history:
-				urllist = RSSBaseView.findEnclosure(self, content[3])
-				if urllist and self.feed and self.feed.title != _("New Items"):
-					picfile = join(TEMPPATH, url2filename(urllist[0]))
-					pngfile = "%s.png" % picfile.split(".")[0]
-					if not exists(pngfile):
-						callInThread(downloadPicfile, urllist[0], picfile, resize=self.picsize, callback=self.refreshPics)
-						pngfile = self.nopic
-				else:
-					pngfile = self.nopic
+				enclist = RSSBaseView.findEnclosure(self, content[3])
+				pngfile = self.nopic
+				if self.feed and self.feed.title != _("New Items"):
+					for enclosure in enclist:
+						if isExtensionSupported(enclosure[0]):
+							picfile = join(TEMPPATH, url2filename(enclosure[0]))
+							pngfile = "%s.png" % picfile[:picfile.rfind(".")]
+							if not exists(pngfile):
+								callInThread(downloadPicfile, enclosure[0], picfile, resize=self.picsize, callback=self.refreshPics)
+								pngfile = self.nopic
+								break
 				picpix = LoadPixmap(cached=True, path=pngfile if exists(pngfile) else join(TEMPPATH, NOPIC))
 				skinlist.append((content[0], self.linepix, picpix))
 			self["content"].updateList(skinlist)
@@ -641,7 +683,11 @@ class RSS_FeedView(RSSBaseView):  # Shows a RSS-Feed
 		skinlist = []
 		if self.feed:
 			for content in self.feed.history:
-				picurl = content[3][0][0] if len(content[3]) > 0 else ""
+				picurl = ""
+				for attachments in content[3] if len(content[3]) > 0 else []:
+					if attachments[1] in MIMEIMAGES:  # catch first supported picture-url in list of attachments
+						picurl = attachments[0]
+						break
 				pngfile = join(TEMPPATH, url2filename(picurl, forcepng=True)) if picurl else self.nopic
 				picpix = LoadPixmap(cached=True, path=pngfile if exists(pngfile) else join(TEMPPATH, NOPIC))
 				skinlist.append((content[0], self.linepix, picpix))
@@ -798,7 +844,7 @@ class RSS_Overview(RSSBaseView):  # Shows an Overview over all RSS-Feeds known t
 			logourl = feed[0].logoUrl
 			if logourl:
 				logofile = join(TEMPPATH, url2filename(logourl))
-				logopng = "%s.png" % logofile.split(".")[0]
+				logopng = "%s.png" % logofile[:logofile.rfind(".")]
 				if not exists(logopng):
 					callInThread(downloadPicfile, logourl, logofile, resize=self.logosize, callback=self.refreshPics)
 			else:
@@ -901,7 +947,7 @@ class RSSPoller:  # Keeps all Feed and takes care of (automatic) updates
 				self.doCallback(ident)
 		except NotImplementedError as errmsg:
 			if ident:  # Don't show this error when updating in background
-				AddPopup(_("Sorry, this type of feed is unsupported:\n%s") % str(errmsg), MessageBox.TYPE_INFO, timeout=3)
+				AddPopup(_("Sorry, this type of feed is unsupported:\n%s") % str(errmsg), type=MessageBox.TYPE_INFO, timeout=3)
 			else:
 				self.next_feed()  # We don't want to stop updating just because one feed is broken
 		except Exception:
@@ -923,15 +969,16 @@ class RSSPoller:  # Keeps all Feed and takes care of (automatic) updates
 				response.raise_for_status()
 				xmlData = response.content
 				response.close()
-				try:
+#				try:
+				if 1 == 1:
 					if xmlData:
 						self.gotPage(xmlData)
 					else:
 						print("[%s] ERROR in module 'pollXml': server access failed, no xml-data found." % MODULE_NAME)
-				except Exception as err:
-					print("[%s] ERROR in module 'pollXml': invalid xml data from server. %s" % (MODULE_NAME, str(err)))
-					if errorback:
-						errorback(str(err))
+#				except Exception as err:
+#					print("[%s] ERROR in module 'pollXml': invalid xml data from server. %s" % (MODULE_NAME, str(err)))
+#					if errorback:
+#						errorback(str(err))
 			except exceptions.RequestException as err:
 				print("[%s] ERROR in module 'pollXml': '%s" % (MODULE_NAME, str(err)))
 				if errorback:
@@ -963,7 +1010,7 @@ class RSSPoller:  # Keeps all Feed and takes care of (automatic) updates
 					RemovePopup(NOTIFICATIONID)
 					AddNotificationWithID(NOTIFICATIONID, RSS_FeedView, feed=self.newItemFeed, newItems=True)
 				elif update_notification_value == "notification":
-					AddPopup(_("Received %d new news item(s).") % (len(self.newItemFeed.history)), MessageBox.TYPE_INFO, 5, NOTIFICATIONID)
+					AddPopup(_("Received %d new news item(s).") % (len(self.newItemFeed.history)), NOTIFICATIONID, type=MessageBox.TYPE_INFO, timeout=5)
 				elif update_notification_value == "ticker":
 					if tickerView:
 						tickerView.display(self.newItemFeed)
@@ -1057,6 +1104,11 @@ class RSSEntryWrapper(ElementWrapper):
 			myl = []
 			for elem in self._element.findall("%senclosure" % self._ns):
 				myl.append((elem.get("url"), elem.get("type"), elem.get("length")))
+			for elem in self._element.findall("%sdescription" % self._ns):  # alternative search for enclosures
+				res = search(r'src="(.*?)"', elem.text)  # perhaps not the most elegant way but it works
+				if res:
+					url = res.group(1)
+					myl.append((url, EXT2MIME.get(url[url.rfind("."):], "unknown"), "0"))
 			return myl
 		elif tag == "id":
 			return self._element.findtext("%sguid" % self._ns, "%s%s" % (self.title, self.link))
@@ -1194,7 +1246,7 @@ class UniversalFeed(BaseFeed):  # Feed which can handle rdf, rss and atom feeds 
 		del self.history[self.MAX_HISTORY_ELEMENTS:]  # Eventually cut history
 		return self.history[:idx]
 
-	def gotFeed(self, feed):  # select wrapping method and
+	def gotFeed(self, feed):  # select wrapping method
 		if self.wrapper:
 			wrapper = self.wrapper(feed, self.ns)
 		else:
@@ -1219,9 +1271,7 @@ class UniversalFeed(BaseFeed):  # Feed which can handle rdf, rss and atom feeds 
 			self.description = strip_readable(wrapper.description) or ""
 			for child in feed.findall("channel"):  # perhaps not the most elegant way but it works
 				image = child.find("image")
-				if image:
-					logoUrl = image.find("url").text
-					self.logoUrl = logoUrl
+				self.logoUrl = image.find("url").text if image else ""
 		return self.gotWrapper(wrapper)
